Refactor login/recover password reducers
